Refactor: extract UploadFixtures to dry test code
